test: use make's integrated test harness
We use the integrated test harness support of GNU make by building the
"check" target. Currently, we comment out xfails that potentially have
deadlocks because there seems to be no way to specify timeouts in the
integrated test harness of GNU make. Moving to other test drivers,
such as TAP and DejaGnu, would solve this.