tlp: radical change in the protocol to follow the same as mcsp
The previous protocol hurts throughput and does not honor properly the
prioritization. The main reason is the "common" lock. It incurs
substantial sharing and thus expensive data movement. Also, high
priority waiters are only detected once "common" is set to REQUEST. The
detection is often late and the throughput of high priority threads is
only marginally over that of low priority threads. Thanks to the
"nowaiters" routine, this patch allows the generic TLP lock to follow
the same protocol as MCSP and improves the throughput considerably,
especially with HMCS.