lock: split the API to context-full and context-less
The original API forced a context-full model on the user, who must pass a local
context, such as a qnode in MCS, to the lock API routines. This model requires
TLS or some dynamic way to manage local nodes that the user might not want to
deal with. This API is still useful when local nodes can be kept on the stack
(e.g., short critical section in the same structured block), which cuts down on
the branches and TLS overheads of the context-less API. This latter is the
new default API; the user is only required to pass the lock object without the
need for a local node. The local nodes are allocated on the heap and indexed
with the thread id, which is stored in TLS. The testsuite was modified the use
the context-less API. MMCS was not updated with this change, so the testsuite
ignores this lock implementation.