mcs: make the user-visible type opaque
The internals of the user-visible type zm_mcs_t was exposed to the user
and contains hwloc symbols. This forces izem users to unecessarily deal
with having the hwloc dependency in their code. This patch makes this
dependency disapear. In fact, many other types should be opaque and
should be fixed in the future