hmpr: first draft
HMPR (Hierarchical MCS with Priority Restriction) is an arbitrary-level loose
priority lock that uses HMCS and restricts threads with priority > 0 (0 being
the highest priority). HMPR has an interface that allows raising a thread's
priority in user space. Threads with priority = 0 proceed to compete for an HMCS
lock. If no thread with priority 0 exists, FIFO among threads with priority > 0
is adopted.