lock: only build lock-related code if hwloc is present
Before this patch, it was either hwloc is set or izem would't be built.
hwloc is only needed for HMCS (MCS could be changed to avoid this
dependency). The right solution is to only avoid building HMCS and
code depending on it (e.g., TLP with HMCS). This patch is a heavy hammer
workaround to allow at least the queue interface usable when hwloc is
not present.