common: added the gcc __sync atomics as the last fallback
Macros were sufficient to map izem atomics to the __sync equivalent except for
the SWAP or (exchange) operation. Since it required using a function instead,
types must be handled manually. This patch does so only for integers and
pointers; more types might be added on the go though we might drop this support
in favor of a gcc >= 4.7.2 compatibility requirement and use the memory aware
__atomic version.