other: added DSM-Sync, a software combining method
DSM-Sync [1] (or Synch as called by the original authors) is a software
combining technique that augments an MCS locks with combining, a method
whereby a thread holding the lock executes work requests on behalf of
threads waiting to enter the critical section.

[1] Fatourou, Panagiota, and Nikolaos D. Kallimanis. "Revisiting the combining
synchronization technique." In ACM SIGPLAN Notices, vol. 47, no. 8, pp.