dsm: augment with traditional acquire/release operations
The user might want to access DSM-Sync-protected objects in a
traditional critical section fashion, in addition to using software
combining. A motivating example is a situation where software combining
would be used for high contention code paths and critical sections for
less contended paths (to avoid unnecessary overheads). The same
object(s) are basically accessed by multiple types of call sites, and
not all of them need to be converted to "apply" operators (function
pointers at the moment) to use software combining, which incurs
additionally branches and load/store operations.