dsm: make the scope of an "apply" operator qnode-level
Before and in the original design of DSM/CC-Sync, the apply operator is
passed to the "sync" API function. The apply operator may be applied
not only to the caller's request but also to any request pulled from the
queue. We refer to the scope of the apply operator in this case as
"queue-level", because the same operator is applied across all nodes in
the queue.

The consequence of this design is that one cannot execute two different
apply operators on the same sync object. For example, if a concurrent
queue is protected by a single sync object, we cannot perform "enq" and
"deq" operators on it, since there are two of them. This patch
eliminates this restriction by embedding the apply operator in the qnode
and invoque it similarily to how a class method is invoked on the object
instanciation. This effictively allows users to pass any number of
distinct apply operators.