idsm: fix under "head" value on lock acquisition
The "release" needs tnode->head to point at the head of the queue. The
traversal usually happens in "combine" but the acquisition inteface
doesn't call it, which breaks the "release" call. This patch fixes it by
setting the head of the queue to my local node, which was used to
acquire the lock.