dsm: fixing bug that overwrites the "next" field on release
This stems from the original algorithm by Fatourou et al. The next field
is unnecessary overwritten by a NULL value while the owner of the head
node might be attempting to enqueue itself. The next lock holder will be
stuck waiting for the next node inline to update its "next" field to a
non-NULL value, but it will never happen since the previous lock holder
overwritten it to NULL.